requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
attrs==24.2.0
Authlib==1.6.0
backoff==2.2.1
bcrypt==4.2.0
bip32==4.0
bip44==0.1.5
bitarray==3.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cloudinary==1.44.1
coincurve==20.0.0
colorama==0.4.6
cryptography==44.0.1
cytoolz==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
filetype==1.2.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
Mako==1.3.9
MarkupSafe==2.1.5
marshmallow==3.23.1
mistune==3.1.2
mnemonic==0.21
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==2.1.3
oauthlib==3.2.2
openai==1.107.0
opencage==3.0.3
opencv-python==4.11.0.86
OWSLib==0.32.0
packaging==24.1
parsimonious==0.10.0
pillow==11.0.0
imagehash==4.3.2
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.1.4
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.25.7
rlp==4.1.0
rpds-py==0.22.3
rsa==4.9
sentinelhub==3.11.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tifffile==2024.9.20
tomli==2.2.1
tomli_w==1.1.0
toolz==1.0.0
tqdm==4.66.5

types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
unittest-xml-reporting==3.2.0
uritemplate==4.1.1
urllib3==2.2.3
utm==0.7.0
web3==7.10.0
websockets==13.1
Werkzeug==3.0.4
yarl==1.18.0
